reorg network requests by entity
